pyproject.toml
[project]
name = "mchem-si-creator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "python-docx>=1.2.0",
]
